Moved statictics routes to user routes file
